Move cluster_subgraph out of xg
