+++ src/cluster.cpp
@@ -1276,6 +1276,32 @@ vector<pair<size_t, size_t>> OrientedDistanceClusterer::pair_clusters(const vect
     
 };
 
+Graph cluster_subgraph(const xg::XG& xg, const Alignment& aln, const vector<vg::MaximalExactMatch>& mems, double expansion) {
+    assert(mems.size());
+    auto& start_mem = mems.front();
+    auto start_pos = make_pos_t(start_mem.nodes.front());
+    auto rev_start_pos = reverse(start_pos, xg.node_length(id(start_pos)));
+    // Even if the MEM is right up against the start of the read, it may not be
+    // part of the best alignment. Make sure to have some padding.
+    // TODO: how much padding?
+    Graph graph;
+    int padding = 1;
+    int get_before = padding + (int)(expansion * (int)(start_mem.begin - aln.sequence().begin()));
+    if (get_before) {
+        graph.MergeFrom(xg.graph_context_id(rev_start_pos, get_before));
+    }
+    for (int i = 0; i < mems.size(); ++i) {
+        auto& mem = mems[i];
+        auto pos = make_pos_t(mem.nodes.front());
+        int get_after = padding + (i+1 == mems.size() ?
+                                   expansion * (int)(aln.sequence().end() - mem.begin)
+                                   : expansion * max(mem.length(), (int)(mems[i+1].end - mem.begin)));
+        graph.MergeFrom(xg.graph_context_id(pos, get_after));
+    }
+    sort_by_id_dedup_and_clean(graph);
+    return graph;
+}
+
 
 }
